xilpm_ng: Fix compile warnings and tidy prototypes/visibility
- Add missing prototypes in domain headers; include headers in xpm_powerdomain.c
- Mark internal functions static; include missing xpm_device_fsm.h
- Fix types/casts and apply masks for apertures; use unsigned suffixes where needed
- Adjust macro guards for XIicPs; narrow XPmRail_Control mode to u8
- Drop unused variables/externs; update copyright years
- No functional change; warning fixes only


Acked-for-series: Nicole Baze <amanda.baze@amd.com>

+++ lib/sw_services/xilpm_ng/src/xilpm_boot/versal_2ve_2vm/xpm_rail.c
@@ -722,7 +722,7 @@ XStatus XPmRail_AdjustVID(XPm_Rail *Rail)
 			continue;
 		}
 
-		Status = XPmRail_Control(Rail, (u8)XPM_POWER_STATE_ON, (Index + 2));
+		Status = XPmRail_Control(Rail, (u8)XPM_POWER_STATE_ON, (u8)(Index + 2));
 		if (XST_SUCCESS == Status) {
 			VIDAdj->VIDAdjusted = 1;
 			break;
